fix: edge tunnel go routine leak
